requirements.txt
streamlit
pdfplumber
pandas
Pillow
